some clean up

+++ src/gui_draw.c
@@ -288,11 +288,11 @@ static void draw_round_slab_sprite(long pos_x, long pos_y, int units_per_px, con
 void draw_round_slab64k_remap(long pos_x, long pos_y, int units_per_px, long width, long height, long style_type, const unsigned char *cmap)
 {
     unsigned short drwflags_mem = RendererGetDrawFlags();
-    long fill_inset = scale_ui_value_lofi(4);
+    int32_t fill_inset = scale_ui_value_lofi(4);
     /* Keep the fill out of the rounded outer edge, but cover the transparent
        interior of the larger corner and button sprites. */
-    long corner_width = scale_ui_value_lofi(12);
-    long corner_height = scale_ui_value_lofi(12);
+    int32_t corner_width = scale_ui_value_lofi(12);
+    int32_t corner_height = scale_ui_value_lofi(12);
     RendererClearDrawFlags(Lb_SPRITE_OUTLINE);
     if (style_type == ROUNDSLAB64K_LIGHT) {
         RendererAddDrawFlags(Lb_SPRITE_TRANSPAR4);
@@ -311,7 +311,7 @@ void draw_round_slab64k_remap(long pos_x, long pos_y, int units_per_px, long wid
     int y;
     const struct TbSprite* spr = get_panel_sprite(GPS_message_frame_thin_hex_ct);
     int ps_units_per_spr = scale_ui_value_lofi(416)/spr->SWidth;
-    long i;
+    int32_t i;
     for (i = 0; i < width - scale_ui_value_lofi(68); i += scale_ui_value_lofi(26))
     {
         x = pos_x + i + scale_ui_value_lofi(34);
